+++ lightningd/Makefile
@@ -3,6 +3,7 @@
 LIGHTNINGD_SRC :=				\
 	lightningd/anchorspend.c		\
 	lightningd/bitcoind.c			\
+	lightningd/broadcast.c			\
 	lightningd/chaintopology.c		\
 	lightningd/watchman.c			\
 	lightningd/channel.c			\
