+++ lightningd/chaintopology.h
@@ -1,6 +1,7 @@
 #ifndef LIGHTNING_LIGHTNINGD_CHAINTOPOLOGY_H
 #define LIGHTNING_LIGHTNINGD_CHAINTOPOLOGY_H
 #include "config.h"
+#include <lightningd/broadcast.h>
 #include <lightningd/watch.h>
 
 struct bitcoin_tx;
@@ -14,20 +15,6 @@ struct wallet;
 
 /* We keep the last three in case there are outliers (for min/max) */
 #define FEE_HISTORY_NUM 3
-
-/* Off topology->outgoing_txs */
-struct outgoing_tx {
-	struct channel *channel;
-	const struct bitcoin_tx *tx;
-	struct bitcoin_txid txid;
-	u32 minblock;
-	bool allowhighfees;
-	const char *cmd_id;
-	bool (*finished)(struct channel *channel, const struct bitcoin_tx *,
-			 bool success, const char *err, void *arg);
-	bool (*refresh)(struct channel *, const struct bitcoin_tx **, void *arg);
-	void *cbarg;
-};
 
 struct block {
 	u32 height;
@@ -69,27 +56,6 @@ static inline bool block_eq(const struct block *b, const struct bitcoin_blkid *k
 }
 HTABLE_DEFINE_NODUPS_TYPE(struct block, keyof_block_map, hash_sha, block_eq,
 			  block_map);
-
-/* Hash blocks by sha */
-static inline const struct bitcoin_txid *keyof_outgoing_tx_map(const struct outgoing_tx *t)
-{
-	return &t->txid;
-}
-
-static inline size_t outgoing_tx_hash_sha(const struct bitcoin_txid *key)
-{
-	size_t ret;
-	memcpy(&ret, key, sizeof(ret));
-	return ret;
-}
-
-static inline bool outgoing_tx_eq(const struct outgoing_tx *b, const struct bitcoin_txid *key)
-{
-	return bitcoin_txid_eq(&b->txid, key);
-}
-HTABLE_DEFINE_DUPS_TYPE(struct outgoing_tx, keyof_outgoing_tx_map,
-			outgoing_tx_hash_sha, outgoing_tx_eq,
-			outgoing_tx_map);
 
 /* Our plugins give us a series of blockcount, feerate pairs. */
 struct feerate_est {
@@ -205,48 +171,6 @@ u32 penalty_feerate(struct chain_topology *topo);
 
 /* Usually we set nLocktime to tip (or recent) like bitcoind does */
 u32 default_locktime(const struct chain_topology *topo);
-
-/**
- * broadcast_tx - Broadcast a single tx, and rebroadcast as reqd (copies tx).
- * @ctx: context: when this is freed, callback/retransmission don't happen.
- * @topo: topology
- * @channel: the channel responsible for this (stop broadcasting if freed).
- * @tx: the transaction
- * @cmd_id: the JSON command id which triggered this (or NULL).
- * @allowhighfees: set to true to override the high-fee checks in the backend.
- * @minblock: minimum block we can send it at (or 0).
- * @finished: if non-NULL, call every time sendrawtransaction returns; if it returns true, don't rebroadcast.
- * @refresh: if non-NULL, callback before re-broadcasting (can replace tx):
- *           if returns false, delete.
- * @cbarg: argument for @finished and @refresh
- */
-#define broadcast_tx(ctx, topo, channel, tx, cmd_id, allowhighfees,	\
-		     minblock, finished, refresh, cbarg)		\
-	broadcast_tx_((ctx), (topo), (channel), (tx), (cmd_id), (allowhighfees), \
-		      (minblock),					\
-		      typesafe_cb_preargs(bool, void *,			\
-					  (finished), (cbarg),		\
-					  struct channel *,		\
-					  const struct bitcoin_tx *,	\
-					  bool, const char *),		\
-		      typesafe_cb_preargs(bool, void *,			\
-					  (refresh), (cbarg),		\
-					  struct channel *,		\
-					  const struct bitcoin_tx **),	\
-		      (cbarg))
-
-void broadcast_tx_(const tal_t *ctx,
-		   struct chain_topology *topo,
-		   struct channel *channel,
-		   const struct bitcoin_tx *tx TAKES,
-		   const char *cmd_id, bool allowhighfees, u32 minblock,
-		   bool (*finished)(struct channel *,
-				    const struct bitcoin_tx *,
-				    bool success,
-				    const char *err,
-				    void *),
-		   bool (*refresh)(struct channel *, const struct bitcoin_tx **, void *),
-		   void *cbarg TAKES);
 
 struct chain_topology *new_topology(struct lightningd *ld, struct logger *log);
 void setup_topology(struct chain_topology *topology);
