+++ lightningd/peer_control.c
@@ -318,7 +318,7 @@ static struct bitcoin_tx *sign_and_send_last(const tal_t *ctx,
 
 	/* Keep broadcasting until we say stop (can fail due to dup,
 	 * if they beat us to the broadcast). */
-	broadcast_tx(channel, ld->topology, channel, tx, cmd_id, false, 0,
+	broadcast_tx(channel, ld, channel, tx, cmd_id, false, 0,
 		     commit_tx_send_finished, NULL, take(adet));
 
 	return tx;
